add popular tags route, hide unnecessary routes

+++ src/modules/projects/projects.controller.ts
@@ -1,4 +1,17 @@
-import { Body, Controller, Delete, Get, HttpStatus, Param, Post, Put, Query, Req, UseGuards } from '@nestjs/common';
+import {
+  Body,
+  Controller,
+  Delete,
+  Get,
+  HttpStatus,
+  Param,
+  Post,
+  Put,
+  Query,
+  Req,
+  UnauthorizedException,
+  UseGuards,
+} from '@nestjs/common';
 import { ApiBearerAuth, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { AccessTokenGuard } from '../../common/guards/access-token.guard';
 import { AuthUserRequest } from '../auth/types/auth-user-request';
@@ -47,7 +60,15 @@ export class ProjectsController {
   @ApiResponse({ status: HttpStatus.OK, description: 'Success', type: Project })
   @ApiResponse({ status: HttpStatus.UNAUTHORIZED, description: 'Unauthorized' })
   @ApiResponse({ status: HttpStatus.NOT_FOUND, description: 'Not Found' })
-  async updateOneById(@Param('key') key: string, @Body() updateDto: UpdateProjectDto): Promise<Project> {
+  async updateOneById(
+    @Req() req: AuthUserRequest,
+    @Param('key') key: string,
+    @Body() updateDto: UpdateProjectDto,
+  ): Promise<Project> {
+    const project = await this.projectsService.findOne('_id', key);
+    if (project.creator._id.toString() !== req.user.sub) {
+      throw new UnauthorizedException('You are not allowed to update this project');
+    }
     return this.projectsService.updateOne('_id', key, updateDto);
   }
 
@@ -67,7 +88,11 @@ export class ProjectsController {
   @ApiResponse({ status: HttpStatus.OK, description: 'Success', type: Project })
   @ApiResponse({ status: HttpStatus.UNAUTHORIZED, description: 'Unauthorized' })
   @ApiResponse({ status: HttpStatus.NOT_FOUND, description: 'Not Found' })
-  async deleteOneById(@Param('key') key: string): Promise<Project> {
+  async deleteOneById(@Req() req: AuthUserRequest, @Param('key') key: string): Promise<Project> {
+    const project = await this.projectsService.findOne('_id', key);
+    if (project.creator._id.toString() !== req.user.sub) {
+      throw new UnauthorizedException('You are not allowed to update this project');
+    }
     return this.projectsService.deleteOne('_id', key);
   }
 }
